+++ miraiSugoroku/src/main/resources/static/Masu.css
@@ -1,51 +1,51 @@
 
 
-section img#masu1{
+section img#masu_start{
     top: 500px;
     left: 100px;
     position: absolute;
 }
-section img#masu2{
+section img#masu1{
     top: 500px;
     left: 300px;
     position: absolute;
 }
-section img#masu3{
+section img#masu2{
     top: 500px;
     left: 500px;
     position: absolute;
 }
-section img#masu4{
+section img#masu3{
     top: 500px;
     left: 700px;
     position: absolute;
 }
-section img#masu5{
+section img#masu4{
     top: 500px;
     left: 900px;
     position: absolute;
 }
-section img#masu6{
+section img#masu5{
     top: 500px;
     left: 1100px;
     position: absolute;
 }
-section img#masu7{
+section img#masu6{
     top: 500px;
     left: 1300px;
     position: absolute;
 }
-section img#masu8{
+section img#masu7{
     top: 500px;
     left: 1500px;
     position: absolute;
 }
-section img#masu9{
+section img#masu8{
     top: 500px;
     left: 1700px;
     position: absolute;
 }
-section img#masu10{
+section img#masu9{
     top: 500px;
     left: 1900px;
     position: absolute;
